% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{assign_gc_bins}
\alias{assign_gc_bins}
\title{assign_gc_bins}
\usage{
assign_gc_bins(gc)
}
\arguments{
\item{gc}{Numeric vector of GC fractions.}
}
\value{
An integer vector of bin indices.
}
\description{
Assign GC values to five bins delimited by their own
quantiles.
}
\keyword{internal}
