% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{calculate_gcdist}
\alias{calculate_gcdist}
\title{calculate_gcdist}
\usage{
calculate_gcdist(genome, threads = 1)
}
\arguments{
\item{genome}{The genome object to use, no default value is set}

\item{threads}{The number of threads to use for parallel processing, default is 1}
}
\value{
GC content values for each chromosome
}
\description{
calculating GC distribution and GC frequency table
}
