% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{compute_gc}
\alias{compute_gc}
\title{compute_gc}
\usage{
compute_gc(X)
}
\arguments{
\item{X}{The DNA sequence to compute GC content for}
}
\value{
GC content value
}
\description{
compute the GC content for a given DNA sequence
}
