% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{convert_to_bins}
\alias{convert_to_bins}
\title{convert_to_bins}
\usage{
convert_to_bins(x, gc_bin)
}
\arguments{
\item{x}{The GC content values to convert}

\item{gc_bin}{The GC bins to use for conversion}
}
\value{
A matrix of bin indices
}
\description{
convert GC content values to bins
}
