% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{convert_to_matrix}
\usage{
convert_to_matrix(bins)
}
\arguments{
\item{bins}{The bin indices to convert}
}
\value{
A matrix of bin indices
}
\description{
convert bin indices to a matrix
}
