% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package.R
\name{createMethylTFRPackageScaffold}
\alias{createMethylTFRPackageScaffold}
\title{createMethylTFRPackageScaffold}
\usage{
createMethylTFRPackageScaffold(
  assembly,
  dest = getwd(),
  motifSets = c("JASPAR2020")
)
}
\arguments{
\item{assembly}{The genome assembly, e.g., "hg38".}

\item{dest}{Destination directory where the package should be created.}

\item{motifSets}{a character vector of motif sets to be included in the package.}
}
\value{
Invisibly, \code{TRUE} if the package directory and its \code{DESCRIPTION} file were successfully created;
        \code{FALSE} otherwise.
}
\description{
Creates a scaffold folder structure for a methylTFR annotation package.
}
\examples{
createMethylTFRPackageScaffold("hg38")
}
\author{
methylTFR Authors
Inspired by \code{createPackageScaffold} from the \code{RnBeadsAnnotationCreator} package.
}
