% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findTFBindSites}
\alias{findTFBindSites}
\title{findTFBindSites}
\usage{
findTFBindSites(genome, motifs, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{genome}{BSgenome object}

\item{motifs}{PWMatrixList or PFMatrixList of motifs}

\item{BPPARAM}{Parallel backend (default: automatically selected by bpparam())}
}
\value{
A GRangesList of binding sites for each motif
}
\description{
Find TF binding sites from a genome and motif list (parallelized across chromosomes)
}
