% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_gcdist}
\alias{get_gcdist}
\title{get_gcdist}
\usage{
get_gcdist(chr, chr_len, genome)
}
\arguments{
\item{chr}{The chromosome name}

\item{chr_len}{The length of the chromosome}

\item{genome}{The genome object to use}
}
\value{
GC content values for the chromosome
}
\description{
get the GC distribution for a given chromosome
}
