% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenomeObject}
\alias{getGenomeObject}
\title{getGenomeObject}
\usage{
getGenomeObject(assembly, adjChrNames = TRUE)
}
\arguments{
\item{assembly}{string specifying the assembly (e.g., "hg19", "hg38")}

\item{adjChrNames}{should chromosome names be adjusted (e.g., adding "chr" prefix)?}
}
\value{
\code{BSgenome} object
}
\description{
Retrieve the appropriate \code{BSgenome} object for a given assembly string
}
\author{
Fabian Mueller
}
