% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepareMotifmatchr}
\alias{prepareMotifmatchr}
\title{prepareMotifmatchr}
\usage{
prepareMotifmatchr(genome, motifs)
}
\arguments{
\item{genome}{character string specifying genome assembly}

\item{motifs}{either a character string ("jaspar", "encode", "cisbp", etc.) or an object containing PWMs}
}
\value{
A list containing objects to be used as arguments for \code{motifmatchr}
}
\description{
Prepare objects for a \code{motifmatchr} analysis
}
\author{
Fabian Mueller
}
