% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{processMotifs2Matrix}
\alias{processMotifs2Matrix}
\title{processMotifs2Matrix}
\usage{
processMotifs2Matrix(motif, gc_bin, genome, tf_bindsites, enhancer = NULL)
}
\arguments{
\item{motif}{The motif to process}

\item{gc_bin}{The GC bins to use for conversion}

\item{genome}{The genome object to use}

\item{tf_bindsites}{The TF binding sites to use}

\item{enhancer}{The enhancer regions to use}
}
\value{
A matrix of GC content values
}
\description{
process motifs to matrix
}
