% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annotation.R
\name{build_annotations}
\alias{build_annotations}
\title{Build Annotations}
\usage{
build_annotations(
  annotations,
  pkg.base.dir,
  chunk_size = 10,
  genome,
  cores = 10,
  enhancer = NULL
)
}
\arguments{
\item{annotations}{Character vector of motif set names (e.g., "cisbp", "encode") OR a named list of GRanges objects}

\item{pkg.base.dir}{Output directory to save RDS files (default: inst/extdata)}

\item{chunk_size}{Number of annotations to save per RDS (default: 10)}

\item{genome}{Character genome assembly (e.g., "hg38") or BSgenome object}

\item{cores}{Number of cores to use for parallel processing (default: 10)}

\item{enhancer}{GRanges object for distal regions (default: NULL)}
}
\description{
Build annotation RDS files from either motif set names or provided GRanges.
}
\author{
Irem Gunduz
}
