% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{computeGCgenome}
\alias{computeGCgenome}
\title{computeGCgenome}
\usage{
computeGCgenome(genome, cores = 1)
}
\arguments{
\item{genome}{The genome object to use}

\item{cores}{The number of cores to use for parallel processing}
}
\value{
A GRanges object with GC content values for the entire genome
}
\description{
compute GC content for the entire genome
}
