% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{computeGCgenome_helper}
\alias{computeGCgenome_helper}
\title{computeGCgenome_helper}
\usage{
computeGCgenome_helper(genome, chr, chr_len)
}
\arguments{
\item{genome}{The genome object to use}

\item{chr}{The chromosome name}

\item{chr_len}{The length of the chromosome}
}
\value{
A GRanges object with GC content values
}
\description{
compute GC content for a given chromosome
}
\keyword{internal}
